test-requirements.txt
flake8>=2.2.0
pytest>=2.8.0
pytest-cov>=2.1.0
PyHamcrest>=1.8.5